Ran regression; Fix expression to dump structures on stack.
 Changes to be committed:
	modified:   regress_report.txt

1. Ran regression.

	modified:   source/pcom.pas

1. Fixed expression to dump one or two structures off the stack, see note.

Note:

So expression is the only expression node that can take a structure, usually
addressed, but it can take a structured return. After they get loaded, they
turn to vars (addressed), which is good, it places an address over the stacked
structure and then all the normal operators can use that. They just need to be
cleaned off stack when the operation is over. So we needed to preserve the
stacked status and do the cleanup after.