run regression; add "undefined structure" pointer.
From the original P4 compiler, there are a set of undefined identifiers, but no
undefined structure entries (types). This resulted in a series of nil checks
P4 effectively uses nil to represent undefined or malformed structures (types).
I don't believe this is the most efficient error handling mechanisim, and makes
it more difficult to deal with type errors. What was added here is a new
structure entry, utypeptr, that is returned for malformed or missing types.

This entry will cause type miscompares to be supressed. Going forward, the check
for utypeptr will be spread around the compiler to reduce error spew. The
current checks for nil are left alone.

This change didn't affect positive testing, but did affect the PRT. It has been
adjusted to account for this.

 Changes to be committed:
	modified:   regress_report.txt

1. Ran regression.

	modified:   source/pcom.pas

1. Create and initialize utypeptr. This basically looks like an integer.

2. Treat any comparision with utypeptr as equal. Note the error that caused
utypeptr will always be output.

	modified:   standard_tests/iso7185prt.cmp

1. Updated.

	modified:   standard_tests/iso7185prt0026.ecp
	modified:   standard_tests/iso7185prt0029.ecp
	modified:   standard_tests/iso7185prt0032.ecp
	modified:   standard_tests/iso7185prt0033.ecp

1. Updated.

	modified:   standard_tests/iso7185prt0033.pas

1. Added "output" to header, this reduces the error count by 1.

	modified:   standard_tests/iso7185prt0036.ecp
	modified:   standard_tests/iso7185prt0038.ecp
	modified:   standard_tests/iso7185prt0504.ecp
	modified:   standard_tests/iso7185prt0511.ecp
	modified:   standard_tests/iso7185prt1900.ecp

1. Updated.