Removed error redirect; Fixed s/b in test; ran regression; reassign error to stderr in cmach.
The obvious thing, assign error to stderr in cmach, caused a difference in the
way cmach operated vs. pint and pmach. The reason is there is no error header
file in gpc standard mode. Error output in stdio is not synced (why would it
be?), so it ends up in random places in the output file. The (unfortunate)
solution was to have cmach emulate pint and pmach for now. This can change later
after bootstrap, when we can get rid of some of the gpc dependencies.

 Changes to be committed:
	modified:   bin/testpascaline

1. Remove (unneeded) error redirects.

	modified:   pascaline_tests/pascaline.pas

1. Fix incorrect s/b statement.

	modified:   regress_report.txt

1. Ran regression.

	modified:   source/cmach/cmach.c

1. Changed all stderr outputs for error back to output.