Add sfr instruction; move mst into routines; refactor framing.
This is part one of a reframing refactor. Now the dummy function result is
placed on stack. Then the parmeters are placed, and the call to routine is
a simple return placed on the stack and call the routine.

The routine itself creates the mark, so mst occurs before ents and ente.
mst now picks up the routine address and places the ra. The next step is
to redo all of the variable and parameter offsets for the new framing.

This push will not pass regression.

 Changes to be committed:
	modified:   source/pcom.pas

1. Separate sfr and mst handlers.

2. Refactor sfr calls from mst calls, sfr goes where the old mst stack
starts were, and the msts get generated at the top of the routines
before ents and ente.

3. The cup instruction is changed to just push the return address and call the
target routine. It no longer finds and sets the mp, that is done in mst now.
Thus it no longer needs the parameter length parameter. It still is passed
to gencupent(), the reason being it needs to know how to adjust the stack after
the call.

	modified:   source/pint.pas

1. Remove the p parameter from the cup instruction.

2. Change mst to pop the return address in the mark structure.

3. Refactor cup to simply push the return address and jump to the target
routine.

Notes:

1. In callnonstandard(), the call to gensfr() is the same for both externl cases
as well as normal. This may need attention.

2. The cuv instruction needs attention.