Fix lip instruction addressing; refactor parameter addressing.
 Changes to be committed:
	modified:   source/pcom.pas

1. Fix lip instruction addressing.

2. Add parmoff() routine to offset all addresses in a parameter list.

3. Set parameter list offset to 0 during formation.

4. Call parmoff() for the parameter list after it is formed at offset.

5. Set locals offset before declare().

Notes:

In the old scheme the parameters are followed by the locals and both are after
the mark. In the new scheme the parameters preceed the mark and the locals. So
we form the parameter list with offset 0 and find the total length. Then the
list addresses are offset from the calculated top of the list.