Fix for procedure/function parameters.
The cip instruction substitutes the mp for the procedure or function parameter
to be called for the caller's mp. This effectively passes the context into the
parameter procedure/function. The issue becomes restoring the correct mp after
the call. This has to be done transparently to the called procedure/fuction. It
does not care or know that it was called as a parameter function/procedure.

We handle this by adding a rip instruction that restores the original mp. The
issue then becomes where to stash the original mp. We do this by recovering the
original mp from the display itself. The runtime code does not know where that
display is, but the compiler does. It has to do that by calculating the offset
where the stack is after the return from the called procedure/function, then
fetching the old mp from that display.

 Changes to be committed:
	modified:   source/pcom.pas

1. Add locspc, locals total space, to procedure/function entries.

2. Add soff, stack offset, variable to callnonstandard().

3. Find soff from the net offset on the stack at the call time (topnew).

4. fix the lpa instruction p=level argument.

5. Generate rip instruction after the cip call in callnonstandard(). This
carries a net offset to the callers display from locals space, function size,
and stack offset at the call time.

6. find and store locals space in procdeclaration().

7. Add rip instruction to tables.

	modified:   source/pint.pas

1. Add rip instructions to instruction table.

2. Fix cip and rip instructions generation.

3. fix cip instruction.

4. Add rip instruction.