Buffer source lines.
The original code from P5 reads directly from the input file prd. This was
modified to read from prd or from an included file (uses or joins), and thus
was encapsulated with (read/eof/eoln) functions.

We change this to operate a line buffer and change the encapsulasation for that.
This has no net effect, but makes the source lines available for annotation.

Note the source lines don't use the last two positions. The end position is a
sentinel for end of line, and the end minus 1 is used to indicate the eoln
active.

Note that we only use one source line buffer, as opposed to a buffer per include
file (as IP pascal does) because the system only bridges files on line
boundaries.

 Changes to be committed:
	modified:   source/pcom.pas

1. Create maxlin, maximum source line size.

2. Create lininx and linbuf types, the buffer for source lines and the index
type.

3. Create srclin, the source line buffer, srcinx, the current source line
character position, and srclen, the size of the input line, which can be zero.

4. Add fileeof(), which wraps the eof status for input types.

5. Refactor all of the encap functions eofinp(), eolninp(), bufinp() and
readinp() to work with linebuffering.

6. Init the system by doing a readline() to start.

	modified:   source/pint.pas