Break out prt command lines for run; Allow --pint parameter in testprog; run regression.
Now that we can pass in options, we need to pass them through the runprt script.
The method I used was to break out the program command lines again, it was
automated by search. This allows individual options to be specified for each
program in an obvious manner.

 Changes to be committed:
	modified:   bin/runprt

1. Create dotestprog() function.

2. Create list of programs and run through the function.

	modified:   bin/testprog

1. Allow --pint (default) option. This allows runprt to always pass an option so
I can fix %^&* bash issues.

	modified:   regress_report.txt

1. Ran regression, but has the same issue, will fix next commit.