Guarded all intermediate output statements; ran regression.
There were a lot of output to intermediate (prr) statements that weren't
qualified by checking the prcode flag. This causes issues with included files
outputting line markers, but regardless of that, it should not happen.

The next part of this fix is to restore the line count on the parent file of the
included file.