Establish special variables in debug mode and read them.
This was a long contemplated change. There are several P-Machine specific values
that are interesting to read, such as pc, sp, mp, etc. There are commands to
read them, but it works better if they are variables, then they can participate
in expressions.

This is one half the fix, it establishes read only.

 Changes to be committed:
	modified:   pascaline_tests/debug_test.inp

	modified:   source/pint.pas

1. Add error for special symbol not found.

2. Add getlab() routine, like getnam(), but only takes alpha characters plus
'_'.

3. Refactored extyp() and all included routines to change sp to syp. This needed
to happen because we use sp in factor and it was shadowed.

4. In factor(), added decode for special variables starting with '@'.