Fix debug test in regression; re-enable joined overload test in pascaline.pas; fix compare files; run regression; fix module overloads.
We add type digests of parameters to all routines. This was necessary because at
the time of parsing the first routine of an overload group, you don't actually
know it is going to be overloaded. This gets fixed in debug by only searching
for the prefix. The type digest is set off by '@', which prevents the user from
aliasing it.

 Changes to be committed:
	modified:   bin/regress

1. Debug test name was incorrect.

	modified:   pascaline_tests/pascaline.pas

1. Re-enable overload tests in joined module.

	modified:   pascaline_tests/pascaline.cmp
	modified:   pascaline_tests/pascaline_cmach.cmp
	modified:   pascaline_tests/pascaline_pmach.cmp

1. Adjust compare files for new test.

	modified:   regress_report.txt

1. Ran regression.

	modified:   source/pcom.pas

1. Rename isovl*() routines to hasovl*() which more accurately states what they
do, namely check if the given function is included in the overload group.

2. Change gencup() to always output type digest.

3. Change block start to always output type digest.

	modified:   source/pint.pas

1. Add bname field to block data. This holds the base name of the routine,
without the type digest appended.

2. Add routine, strassvfl(), that assigns fixed to variable string, but only up
to a certain length.

3. strassvf() is refactored in terms of strassvfl().

4. In generate(), block header definition, find the type digest, if it exists,
on the block name and stores only the base name in bname.

5. In fndrot(), only search base names for routines.

Notes:

This does not fix debug mode for the ability to work with overload groups. I
think this needs a lot of work, like detecting if a given routine is overloaded,
and a system to enable selecting the overload simpler than specifying the whole
digest.