Partial fix for #59: single character constant passed as container.
 Changes to be committed:
 	modified:   pascaline_tests/pascaline.pas

 1. Added test for single character constant to container.

	modified:   pascaline_tests/pascaline.cmp
	modified:   pascaline_tests/pascaline_cmach.cmp
	modified:   pascaline_tests/pascaline_pmach.cmp

1. Updated all compares, but see note on cmach.

	modified:   regress_report.txt

1. Ran regression, one failure.

	modified:   source/pcom.pas

1. In callnonstandard, change matching check to include constant single
character matches container rule.

2. Move parameter check up and add single character rule, plus put the integer
matches real rule. This didn't need to happen before because the code changes
integer to real.

3. Add special case for single character constant pass to container. It is
loaded, copied to addressable buffer, then passed there.

Notes:

There is an issue with pascaline test fh8 on cmach. This was discovered by
mishandling the cmach compare file, so this mistake was left in.