Reordered the mark and saved MP parameters.
The intel enter instruction wants a different order of the saved mp and display.
I know the standard advice is that enter is slower (than just regular
instructions). However, that is for C programs. It gets more useful for fully
nested procedures. In any case it gives more compact code.

 Changes to be committed:
	modified:   ../regress_report.txt

1. Ran regression. There is a problem with the debugger test, will fix in next
commit.

	modified:   ../source/cmach/mpb64.inc
	modified:   ../source/mpb64.inc

1. Changed mark offsets.

	modified:   ../source/cmach/cmach.c
	modified:   ../source/pint.pas
	modified:   ../source/pmach.pas

1. The mst instruction was reordered.

2. The returns were refactored for the new framing order.

	modified:   ../source/pgen_gcc_AMD64.pas

1. Development on the mst instruction.