Fix parlab to advance over its numeric parameter; fix up mst instruction.
 Changes to be committed:
	modified:   source/pgen_gcc_AMD64.pas

1. parlab() needs to terminate after parsing a number. This fix was in pint.

2. Refactor mst instruction to use enter, but not use the locals allocate from
enter. I was going to, but it doesn't clear memory. See note.

Note:

We can optimize what to clear or not clear later. Files need it, pointers self
check with it. We might be able to come up with a scheme for individual
assignments later. Until then we clear everything, globals, locals and dynamic.