Remove unused variable; Add parameter count/remove.
Have to count parameters in user call. In pint/pmach/cmach, parameter remove is
done by the callee, and the caller does not have to know what the number of
parameters are. In generated code, that's not true. We use a trick, the stack
count at sfr instruction is saved, then at call time we remove the parameters.
If there is no sfr, the parameter count stays invalid and so is ignored. This
is true of things like the program block.

 Changes to be committed:
	modified:   source/pgen_gcc_AMD64.pas

1. Remove unused totreg.

2. Add parlvl counter.

3. At sfr, we activate the parameter level.

4. At cup, we remove stacked parameters until the parameter strike.