Back out last change; add current mp push.
 Changes to be committed:
	modified:   source/pcom.pas

1. Back out locals to mst change. I was wrong, the locals is not a number, it is
a stack position as offset from mp.

	modified:   source/pgen_gcc_AMD64.pas

1. Add push of present mp to mst sequence. This then emulates pint.

Notes:

The framing for Pascal-P6 includes the current mp as part of the display, which
the "enter" instruction was not designed to do. This is for two reasons: First,
its backwards compatible with the framing system, since even the top block has a
display entry, and second, because pcom actually does generate at least one
local access no matter what the block (the for loop index). Yes, I know that it
should not, but again, I like that the system is compatible up and down.