Ran regression; added cuf operator; AMD64 development.
pgen needed a way to differentiate procedurs and functions. This is done in
pint/pmach/cmach at runtime. I added cuf for "call user function". cup then
becomes "call user procedure". Everywhere in pint/pmach/cmach cuf is a proper
alias for cup, because the processing for it is identical outside of pgen.

I tried to use sfr to do this, but sfr has cases where the offset must occur
later in the program, so it was not suitable.

 Changes to be committed:
	modified:   bin/pe

	modified:   regress_report.txt

1. Ran regression, passes complete.

	modified:   source/pcom.pas

1. Add new instruction "cuf".

2. Rename gencup() to gencupcuf().

3. Changed cup to cuf in function calls tro gencupcur().

	modified:   source/pint.pas

1. Added cuf to instruction table.

2. Process cuf as alias to cup.

	modified:   source/cmach/cmach.c
	modified:   source/pmach.pas

1. Process cuf as alias to cup.

	modified:   source/pgen_gcc_AMD64.pas

1. Development.