Ran regression; added step over abilities (second part).
There are two ways to run forward in the code from the debugger. One is to just
exit the debug routine and let the machine run. The other is to run a simulation
loop within the debugger. The difference is basically moot, since the machine
simulator outside of the debugger runs a loop much like the internal loop of the
debug routine.

 Changes to be committed:
	modified:   regress_report.txt

1. Ran regression.

	modified:   source/pint.pas

1. Added setbrk() routine, to set up a breakpoint.

2. Refactor code to use the setbrk() routine.

3. Added skpovr() routine. It acts as a single instruction execute, but checks
if a call instruction is to be executed, then runs forward to the return address
if that is found.

4. Use skpovr() instruction in step routines for step over.