Add dump stack command.
Dump stack is just a typeless and blind command. It does not know what is on the
stack or how many words of stack are in use. This is possible, but would be hard
to arrange. Now it just dumps words off the stack, to a specified depth or 10 if
not specified. It dumps in hex and decimal.

 Changes to be committed:
	modified:   source/pint.pas

1. Add dst command to enum and command table.

2. Add dst code.