Refactor mst instruction.
Besides being incorrect, mst generation used offsets to form the the locals and
stack maximum extents. I changed the mst to reflex the actual amount of space in
locals and the actual space needed for the locals stack. This makes the
accounting a bit easier to reconcille, and it dosen't carry a penalty. Its just
a subtract instead of an add in mst, and use of the sp instead of the mp.

 Changes to be committed:
	modified:   doc/the_p6_compiler.docx

1. I don't recall the change (if any, the editor generates changes for nothing).

	modified:   pascaline_tests/debug_test.cmp

1. Numbers corrections only for debug test.

	modified:   regress_report.txt

1. Ran regression.

	modified:   source/pcom.pas

1. Fix locals parameter for mst instructions, and change to value instead of
offsets.

2. Change local stack offset to positive.

	modified:   source/cmach/cmach.c
	modified:   source/pint.pas
	modified:   source/pmach.pas

1. Refactor mst instruction to take values and not offsets.