Refactor calls.
The previous method was to use the parameter list on the exec stack, but
standard calls moved to a system using a .pl link from expptr entries.
The two methods to pull that off the stack are parameter counts for system
calls, and sfr bracketing for user calls. Refactored in general to use the
new system.

 Changes to be committed:
	modified:   source/pgen_gcc_AMD64.pas

1. Added other forks to deltre().

2. Refactored asspar().

3. Refactored csp, cuf and cup assreg() handlers to use the new asspar() and use
the parameters list.

4. Added callsp() to genexp() that generates code for system calls in tree. Only
system functions are processed this way, but I am going to refactor terminals as
well.

5. Add new routine, getparn(), that takes a number of parameters off the exec
stack and moves them to a parameter list.

6. Refactor getpar() to use getparn().

7. Refactor assemble/csp to use getparn().