Output procedure/function parameter type spagetti; translate '(' and ')' to '-' in assembler; fix cip instruction.
procedure and function parameters were left undefined in type digests. We add the form:

q(p):r

Which is q for procedure function, the parameters in parens, then ':' and the
result type if a function.

GAS cannot take '(' anmd ')' characters. I translate them to '_'. Recall they
are for human consumption, so they are annoying but not wrong.

 Changes to be committed:
	modified:   source/pcom.pas

1. Change wrtsym() to output procedure/function parameters.

2. Change prtpartyp() to output procedure/function parameters.

	modified:   source/pgen_gcc_AMD64.pas

1. Translate '(' and ')' to '_' on input.

2. Fix genexp(): cip.

Note:

I still don't have the indirect call format correct for cip, will fix.