Add insr array, size of instruction stack result; add parreg array, parameter registers.
To add the ability to allocate double registers, I added insr[] array that
contains the number of stack words in the result. Be aware that not all
instructions have an exact result. Also, the current use of this array is to
tell how many parameter registers to allocate, which limits the set of
instructions that would be used for calculation.

 Changes to be committed:
	modified:   source/pgen_gcc_AMD64.pas

1. Added insr array.

2. Added parreg array. This array is 12 locations, 6*2, to cover the available
parameter registers. Its used to allocate registers for parameters. Locations
that cannot be mapped to registers are null. So the maximum possible registers
to come from the array are 6*2, if every parameter was double register.

3. Fill out the insr array.

4. Fill out the parreg[] array.

5. Remove unneeded register reservation in lpa instruction.

Note

This is a partial commit to the coming parameter double register allocation fix.