Refactor the cks/ckv/cke instruction set.
The cks/ckv/cke sequence checks a tag field setting to multiple tag field
constants. The implementation was dependent on the stack. I changed the way it
got implemented to loading the tag into a register, then doing the compare
series in that register. Thus the whole thing becomes a encode led by chke,
the terminal operator. It lets the tag field load, then generates the series
of checks. It does this by removing all ckv checks from the stack and chaining
them, then loading the tag field, then running the individual compares.

 Changes to be committed:
	modified:   doc/the_p6_compiler.docx

1. No change, this is a libreoffice bug.

	modified:   source/pgen_gcc_AMD64.pas

1. Add temp reg to assemble.

2. ckv instructions don't assign registers (no-op).

3. cks just becomes a placeholder.

4. Redid several places where q was referenced where ep^.q should have been.

5. Encode for genexp(): ckv.

6. Change assemble(): ckv to just stack the entries.

7. Refactor cke as described.