Refactor cks/ckv/cke sequences.
The cks/ckv/cke sequences are executed before a load, and they have to be
performed with the expressions they belong to. Thus we move from the terminal
area and refactor extensively.

 Changes to be committed:
	modified:   source/pgen_gcc_AMD64.pas

1. Add cl link to expressions, tag check.

2. Add cl cleanout to deltre().

3. Add cl dump to dmptre().

4. Add register assignment for cke.

5. Add generation for cke.

6. Add attach() routine, takes an expresion entry and attaches any tag check
sequence to that. Used on loads only.

7. Add attach() calls to various load leaves.

8. Add leaf version of cke operator.

9. Remove tree version of cke operator.

Note:

Still need to arrange side calling for cke chains on assreg() and genreg().