Ran regression; Remove options from fpc run; set default for fpc integer to 64 bit.
 Changes to be committed:
	modified:   regress_report.txt

1. Ran regression.

	modified:   source/pcom.pas
	modified:   source/pint.pas

1. Remove options from fpc run.

2. Set integer to 64 bit if in 64 bit mode and FPC is the compiler.

	modified:   source/pmach.pas

1. Remove options from fpc run.

Notes:

FPC uses the C "compatible integer" plan by using 32 bit integers in 64 bit
mode. I changed that model because:

1. That is against the design of Pascal, which specifies integer as the normal
size of the underlying machine and allows the program to adapt (hence maxint).
It is, by the way, against the design of C as well, but that is another story.

2. It wasn't consistently applied. If it were, integer would be 16 bit, since
that was the original word size.

3. It is easier.

4. There is no savings in time for 32 bit vs. 64 bit. Space is another story.