Refactor nwl/dsl handling; Refactor csp/cup register assignment; fix lca error.
nwl and dsl, the new and dispose for variants, treated the stack tag list as a separate
list on the stack. This was refactored into being handled by the general parameter
handling system as a combination of registered and stacked parameters, just as any
call.

 Changes to be committed:
	modified:   source/pgen_gcc_AMD64.pas

1. Remove ndl link.

2. Refactor register overflow checks in asspar().

3. Fix lca register allocate.

4. Refactor csp new/dsl handling. The registered parameters are skipped,
then the rest of the list is processed on stack.

5. getparn() stops at end of stack. This means that nwl/dsl can use it for all
of the tag parameters.

6. Refactor csp as stated to treat nwl/dsl tags as parameters.