Added system to order requested registers; fix register reservations.
I am unhappy with the system of allocating the registers on getreg() in declaration
order. I created a system that gives the order of allocation, and I set this to
the integer order rbx, rcx, rsi, rdi, then registers 8-15. The idea is to allocate
the registers with the least special purposes and that don't need rex extension
codes first. Thus rbx is a freebe, rcx is only used for block moves, as is rsi and
rdi. Then the r8-15, which are orthogonal but need rex codes.

These are tablized, so the order can be rejiggered later. The float registers are
allocated xmm8-15, then xmm0-7, so that the parameter registers are allocated last.

 Changes to be committed:
	modified:   source/pgen_gcc_AMD64.pas

1. Added maxintreg and maxfltreg, giving maximum allocations of registers.

2. Added maxintparreg and maxfltparreg, giving total assignable registers.

3. Added intassord and fltassord, arrays giving the order desired for register
allocation.

4. getreg() and getfreg() are refactored to use the new assignment ordering arrays.

5. resreg() has a quirk that the call order matters. If you reserve registers before
processing forks of the assignment tree, you can cause branches to guard registers.
On the theory that the main use of resreg() is to mark distroyed registers, and that
this applies to the operator itself, I moved the calls to resreg() to the end of
the assreg() routine. Thus the branches of the operator won't see it.