Added double free tracking; Refactor csp file parameter handling; Added error checks.
csp operations that use a file parameter and chain multiple operations using the same
file were refactored. Now they just make a copy of the file access and push it back
onto the stack, then it is dumped at the end of the chain. This is basically part
three of the dup solution.

 Changes to be committed:
	modified:   source/AMD64/gcc/pgen.pas

1. Added sn and free fields to expptr.

2. Check double free in getexp() and putexp().

3. Refactor callsp for new file handling.

4. Dump now just dumps the tree from stack.

5. vre was mislabeled.

6. cjp refactored.

7. csp refactored.

	modified:   source/AMD64/gcc/psystem.c