Add reservation to equ/neq/...r so that left carries to right.
Most of the register allocation cases for dual operands rely on
having the left register assigned and allocated. This happens even
if the result type is different than the operands, like equi.
In the case of real, the demand assign is to an integer register,
and the two operands are float. Thus, in this one case, the left
operand is assigned, then reserved. This effectively communicates
the left operand to the right operand.

The lesson here is: the register allocation in pgen is the minimum
it can be, with relatively primitive operations. Thus it is not always
easy to see how it works in some cases. However, there is usually
a way to steer the system to solve issues, like this one, or
multiply, which has to steer the result around clobbered registers
on the x86 model.