Added dstreg(); change many resreg() instances to dstreg(); move dstreg calls up; add dstreg() calls to functions and procedures.
Added dstreg(). This is a lot alike resreg(), except that it only deals with
reservations on entry. It tells nodes outside of the current branch that the
register will be distroyed, and adds to the push mask if the outside register
is in use.

resreg() also does this, but it flags the register is in use in the node, which
makes sense in the context within the node. dstreg() does not affect the
free registers inside the node, because the register is destroyed but otherwise
is free to use.

 Changes to be committed:
	modified:   source/AMD64/gcc/pgen.pas

1. Added rfs, rf save, to assreg().

2. Added dstreg() routine.

3. Changed resreg() to dstreg() where appropriate.

4. Moved dstreg() calls to the top of assignment routines where possible, but
this was due to a programming error and was not really required. It does tend
to emphasise that the reservation is for node external use, so there you are.

5. Perform a set of dstreg() calls for all of the calling convention registers
that are specified as not preserved on all of csp, cuf, cup, cip and cif.