Fix gettmp() and puttmp(); bridge set address on chks, dif, int and uni; fix various bad set address formulas; add puttmp()s; fix tmpoff start.
 Changes to be committed:
	modified:   source/AMD64/gcc/pgen.pas

1. Fix search bugs in gettmp() and puttmp (lists are hard!).

2. Bridge the set temp addresses on chks, dif, int and uni. The temp addresses
start with the loaders, and must be propagated to the parent nodes. On combining
nodes like dif, int and uni, we have to drop and put one of the addresses and
keep the other branch address. This work is not complete, and all set involving
operators need examination.

3. Remove '$' from lea offset loads. The '^' is becoming the universal escape
in wrtins() macros.

4. In mst, start tmpoff below the display instead of zero.

Notes:

1. Next is to reformulate the locals space in mst incorporating the temps space.

2. I suspect this temp mech is going to be the basis for Pascaline constructs
at a later time.