Added diagnostics for assreg(); fixed diagnostics for rf and rs; created negative diag for rs; fixed sgs.
I figured out that negating rf so that it is "not free registers" will make
it palatible in listings. Thus this and rf dumps were regularized and included in
assreg() and genexp(). Why? Because tracking register use/protection is getting
harder as the code becomes more complex.

Is the output listing too verbose? Yes. Remember you can strip it by eliminating
lines beginning with '#'.

 Changes to be committed:
	modified:   source/AMD64/gcc/pgen.pas

1. Refactor wrtregs() to accept a negation argument. Now, true means dump registers
in set, false means dump registers NOT in set.

2. Added rs (save registers) into dmpety().

3. Make diagnostic for assreg() permanent.

4. Add rf to assreg() dump.

5. Fix sgs.