Rename genexp():pshpar() to genexp():pshparr(); remove assmble():pshpar().
assemble():pshpar() was only used for terminals, which calls no longer are.
Even procedures are processed as expression trees.

pshpar() was renamed to pshparr() because it processes the parameter list
in reverse (right to left), and turns out that is wrong for user calls, since
we reverse the parameter list to order. During this refactor, we thus
classify the pshpars according to type. pshparr() may disappear at some point.