Fix loading of set parameters.
The Pascal-P machine loads sets as values on the stack regardless of length.
We don't do this in pgen, but use a temps system. This works until a call
is made with a set parameter, when the called routine expects the set to
*really* be on stack and essentially turns it into a local. This is because
Pascal-P does not treat sets as structures, and thus does not carry address
for them.

We solve that by hoisting the set parameter onto the stack when encountered.
There is still a registered address for that, which pgen presently does not
use, but the called routine might (C/C++).

Note that we needed to know what the return type of instructions are, and
this was done like floats.

 Changes to be committed:
	modified:   source/AMD64/gcc/pgen.pas

1. Added inss field to instruction map, instruction is set.

2. Added push code to pshpar().