Fix lods; compensate for ipj offsets; reset stkadr.
The 16 byte alignment alignment for stack can affect ipj offset
calculations, so we to a trick. When the stack is aligned, it may
or may not go down 8 bytes. The ipj expects the mark pointer to exist
at that stack location. We write rbp out to the stack after alignment.
If the stack goes down 8 bytes, the rbp is written in duplicate at that new
location. Otherwise it just overwrites the existing copy of rbp and so
is harmless. We can't know if we did the alignment without more work,
so this system removes the need to do that extra work.

Note that ipj offsets need to be compensated for temps use. This fix is
forthcoming.