Added "real register" parameter to assreg().
There is an ambiguous case that we don't know the type of return
from functions. This is a problem in the case of functions returning
real, because we need them to be in the xmm0 register. The answer
chosen was to add a "real register" parameter so that the encoding
node knows to return the result in xmm0 if it is a function. Simply
specifying a real register for the result dosen't work, since in
most cases any floating point register will do (outside of functions).
In this way, the system indicates the needed type of function results.

This is a two part fix.