Fix real return issue; fix rip issue.
There is an ambiguous case that we don't know the type of return
from functions. This is a problem in the case of functions returning
real, because we need them to be in the xmm0 register. To fix this,
pcom.pas checks if the return type of cif/cuf is real, and appends a
flag to cif/cuf instructions that indicates 0 for not real, and 1 for
is a real. Note that this flag is past the end of the instruction and
will be ignored in the interpreter.

I also refactored the rip solution, which was wrong, so that it lets
the cif instruction do it or executes the rip for cip.

 Changes to be committed:
	modified:   source/pcom.pas

1. In gencupcuf(), generate a flag that says if result is real or not.

2. Create gencif() routine that generates a flag for real.

	modified:   source/AMD64/gcc/pgen.pas

1. In assreg():cif, refactor result assignment for real return values.

2. In generate():cip/cif, on cif only, break out the return cases
into rax and xmm0.

3. In generate():cip/cif, on cif only, generate a rip sequence.