Refactor rip; refactor cif/cip for new system; fix strs instruction.
Previous rip fix didn't work. New system is just to ignore rip
instructions and save and restore rbp from rbx instead. rbx is
preserved across calls. The SC in Pascal-P couldn't do this trick
and stay recursive. This is more effient in machine code and removes
the need to track stack in the rip instruction.