Refactor mod and dvi; sign extend for mod.
The idiv instruction on the AMD64 is the most challenging because
both the inputs and outputs are fixed registers, as well as having
side effect registers.

The method I used was to destroy the rax and rdx registers on entry,
then use steering to try to get the rdx and rax registers as
applicable.