Added terminator diagnostic to pgen/genexp(); added intermediate annotation to pgen.
 Changes to be committed:
	modified:   source/AMD64/gcc/pgen.pas

1. Added terminator diagnostic to genexp().

	modified:   source/pcom.pas

1. Added intmsg(), prints intermediate annotation by code.

2. Added digits(), calculates size of printed number.

3. Added par1() and par2() that print and format 1 or 2 numeric parameters.

4. Updated sundry generator routines to do annotation.

Notes:

I have thought about annotation for a long time. This is the least
expensive way, since it only goes in pgen, and carries through to downstream
consumers. It does not decode all the type endings, but those are regular,
so we just output a (t) type note to annotation.

Annotation should help downstream debugging. Hopefully it makes constructing
a generator simpler, since we need lots of them.

The formatting is still pretty rough, but will improve.