Refactored case value errors.
The interpreter uses ujc for case error entries. In the interpreter,
these are the same size as jumps. But in code, they were translated
a formal error call, which was the wrong size.

The solution was to place a code strip in the psystem_asm.s module
and jump to that. The _asm.s module was formerly unused (its for
double length math routines and such). Thus it was brought online.

 Changes to be committed:
	modified:   bin/compile

1. Added psystem_asm.s module to pgen compile.

	modified:   regress_report.txt
	modified:   source/AMD64/gcc/pgen.pas

1. Changed ujc to jump to code strip.

	modified:   source/AMD64/gcc/psystem_asm.s

2. Removed psystem_base() routine, unused. This was done away with
in the framing fix.

	modified:   standard_tests/iso7185prt.cmp
	modified:   standard_tests/iso7185prt0037.ecp

1. Updated the compare files.