Refactor wrtins(); add some annotation to assembly.
 Changes to be committed:
	modified:   regress_report.txt
	modified:   source/AMD64/gcc/pgen.pas

1. Add opcspc, parspc and cmtspc definitions, opcode, parameters and comment
spacings.

2. Move digits() up.

3. Extend instruction string definitons to 60 characters, so added 40, 50 and
60.

4. Added regl() and bregl() functions, returns the length of regular and byte
register strings for formatting.

5. wrtint40() changed to wrtint60().

6. wrtins() was extensively refactored to support
"label opcode parameter comment" format.

7. Added wrtins40() and wrtins50().

8. Instruction defintions everywhere changed to new format. Usually
involved adding a space to the start to indicate opcode is coming, and
adjusted the calls accordingly. Note that the system was not able to
pass labels correctly before.

9. Added test annotation to mst.

Notes:

Will complete the annotations for assembly in the next pass.