Add ability to show call procedure/functions as their actual symbols.
We count intermediates to determine if two labels point to the same location.
When the compiler label is resolved in update(), we search the block list for
a matching block by intermediate count, and mark that link in the compiler
label list. On label searches, that link is returned can can be optionally used.
The only place it is used is cup/cuf, since that is the only place that a
non-compiler label is presented.

Note that the call label only needs to be unique, so we use the abbreviated
overloads symbol system. In this system all of the overloads get numbered by
their encounter order, and the #1 entry does not need a type mark at all.

 Changes to be committed:
	modified:   regress_report.txt
	modified:   source/AMD64/gcc/pgen.pas

1. Add parent link to pblock.

2. Add intermediate count to pblock.

3. Add global intcnt, count intermediates.

4. Add block link to label entries.

5. Add block link to expression entries.

6. Added fndblk(), searches all active blocks for a matching intermediate
count.

7. update() searches for a matching block and sets that in the label table.

8. Add wrtblksht(), write short block name. This uses the encounter order
number along with the short or "typeless" name of the block. If the encounter
number is 1, we drop the whole type ending.

9. Add wrtblktyp(), writes a complete block reference from top to bottom using
short names.

10. Track parent links in generate(): block. This is required because we need
that information even after the block is discarded.

11. Set intermediate count in generate(): block.

12. Label search now returns the block link or nil if there is none.

13. cup/cuf uses the block reference system if available, otherwise default
to compiler labels.

14. Count intermediates in assemble().

15. Expand all labelsearch() references with dummy block link.

Notes:

1. Does not regress.
2. Still need some formatting fixes.